Change the way we handle players playing two games
Rather than adding "duplicate players", use a list rather than a set
to track the players to use to seed games. This makes the checking
for playing games against themselves very easy. To allow for players
to play two games, change seed_games() to accept an optional set of
players who will play two games that round.
Overall, the resulting code is both smaller and simpler. The one new
limitation is that you can't have players play more than two games,
but in practice that seems extremely unlikely to be needed.
This change does also modify the tests, but there may be more tests
that need to be added. In particular, it doesn't add any tests of
the new parameter to seed_games().